+++ static/app/views/replays/detail/accessibility/useAccessibilityFilters.tsx
@@ -3,8 +3,8 @@ import {useCallback, useMemo} from 'react';
 import type {SelectOption} from 'sentry/components/compactSelect';
 import {decodeList, decodeScalar} from 'sentry/utils/queryString';
 import type {HydratedA11yFrame} from 'sentry/utils/replays/hydrateA11yFrame';
-import useSetQueryFieldInLocation from 'sentry/utils/url/useFiltersInLocationQuery';
 import useLocationQuery from 'sentry/utils/url/useLocationQuery';
+import useSetQueryFieldInLocation from 'sentry/utils/url/useSetQueryFieldInLocation';
 import {filterItems} from 'sentry/views/replays/detail/utils';
 
 export interface AccessibilitySelectOption extends SelectOption<string> {
